+++ R/cpi_inflator.R
@@ -11,6 +11,10 @@
 #' The internal data was updated on 2018-08-10 to 2018-Q2. 
 #' If using \code{useABSConnection = TRUE}, ensure you have \code{rsdmx (>= 0.5-10)} up-to-date.
 #' @param allow.projection Should projections beyond the ABS's data be allowed?
+#' @param accelerate.above An integer setting the threshold for 'acceleration'. 
+#' When the maximum length of the arguments exceeds this value, calculate each unique value individually 
+#' then combine. Set to 100,000 as a rule of thumb beyond which calculation speeds benefit
+#' dramatically. Can be set to \code{Inf} to disable acceleration.
 #' @examples 
 #' cpi_inflator(100, from_fy = "2005-06", to_fy = "2014-15")
 #' @return The value of \code{from_nominal_price} in real (\code{to_fy}) dollars.
@@ -20,11 +24,12 @@ cpi_inflator <- function(from_nominal_price = 1,
                          to_fy = NULL, 
                          adjustment = c("seasonal", "none", "trimmed.mean"),
                          useABSConnection = FALSE,
-                         allow.projection = TRUE) {
+                         allow.projection = TRUE,
+                         accelerate.above = 1e5L) {
   # CRAN
-  obsTime <- obsValue <- to_index <- from_index <- NULL
+  obsTime <- obsValue <- NULL
   
-  if (is.null(from_fy) && is.null(to_fy)){
+  if (is.null(from_fy) && is.null(to_fy)) {
     to_fy <- date2fy(Sys.Date())
     from_fy <- prev_fy(to_fy)
     warning("`from_fy` and `to_fy` are missing, using previous and current financial years respectively")
@@ -49,62 +54,66 @@ cpi_inflator <- function(from_nominal_price = 1,
   }
   
   adjustment <- match.arg(adjustment, several.ok = FALSE)
+
   
   
-  
-  
-  if (max.length > 1e5L && 
+  if (max.length > accelerate.above && 
       # don't connect for every group
       !useABSConnection &&
       length(from_nominal_price) == 1L) {
     if (length(to_fy) == 1L) {
-      return(accel_repetitive_input(from_fy,
-                                    cpi_inflator,
-                                    from_nominal_price = from_nominal_price[[1L]],
-                                    to_fy = to_fy[[1L]],
-                                    adjustment = adjustment[[1L]], 
-                                    useABSConnection = FALSE, 
-                                    allow.projection = allow.projection[[1L]]))
-    } else {
+      cpi_fun <- function(x) {
+        cpi_inflator(from_nominal_price = from_nominal_price[[1L]],
+                     from_fy = x,
+                     to_fy = to_fy[[1L]],
+                     adjustment = adjustment[[1L]], 
+                     useABSConnection = FALSE,
+                     allow.projection = allow.projection[[1L]],
+                     accelerate.above = Inf)
+      }
+      return(accel_repetitive_input(from_fy, cpi_fun))
+    }
+    
+    if (length(from_fy) == 1L) {
       cpi_fun <- function(x) {
         cpi_inflator(from_nominal_price = from_nominal_price[[1L]],
                      from_fy = from_fy[[1L]], 
                      to_fy = x,
                      adjustment = adjustment[[1L]], 
                      useABSConnection = FALSE,
-                     allow.projection = allow.projection[[1L]])
+                     allow.projection = allow.projection[[1L]],
+                     accelerate.above = Inf)
       }
       return(accel_repetitive_input(to_fy, cpi_fun))
     }
   }
   
   
-  
-  if (useABSConnection) {
-    switch(adjustment, 
-           "none" = url <-     
-             "http://stat.data.abs.gov.au/restsdmx/sdmx.ashx/GetData/CPI/1.50.10001.10.Q/ABS?startTime=1948", 
-           
-           "seasonal" = url <- 
-             "http://stat.data.abs.gov.au/restsdmx/sdmx.ashx/GetData/CPI/1.50.999901.10+20.Q/ABS?startTime=1948",
-           
-           "trimmed.mean" = url <- 
-             "http://stat.data.abs.gov.au/restsdmx/sdmx.ashx/GetData/CPI/1.50.999902.10+20.Q/ABS?startTime=1948")
-    
-    cpi <- rsdmx::readSDMX(url)
-    message("Using ABS sdmx connection")
-    cpi <- as.data.frame(cpi)
-  } else {
-    switch(adjustment, 
-           "none" = cpi <- cpi_unadj,
-           "seasonal" = cpi <- cpi_seasonal_adjustment,
-           "trimmed.mean" = cpi <- cpi_trimmed)
-  }
-  
   cpi.indices <- 
-    as.data.table(cpi) %>%
-    .[endsWith(obsTime, "Q1")] %>%
-    .[, fy_year := yr2fy(as.integer(sub("-Q1", "", obsTime, fixed = TRUE)))]
+    if (useABSConnection) {
+      switch(adjustment, 
+             "none" = url <-     
+               "http://stat.data.abs.gov.au/restsdmx/sdmx.ashx/GetData/CPI/1.50.10001.10.Q/ABS?startTime=1948", 
+             
+             "seasonal" = url <- 
+               "http://stat.data.abs.gov.au/restsdmx/sdmx.ashx/GetData/CPI/1.50.999901.10+20.Q/ABS?startTime=1948",
+             
+             "trimmed.mean" = url <- 
+               "http://stat.data.abs.gov.au/restsdmx/sdmx.ashx/GetData/CPI/1.50.999902.10+20.Q/ABS?startTime=1948")
+      
+      cpi <- rsdmx::readSDMX(url)
+      message("Using ABS sdmx connection")
+      as.data.frame(cpi) %>%
+        as.data.table(cpi) %>%
+        .[endsWith(obsTime, "Q1")] %>%
+        .[, fy_year := yr2fy(as.integer(sub("-Q1", "", obsTime, fixed = TRUE)))]
+    } else {
+      switch(adjustment, 
+             "none" = cpi_unadj_fy,
+             "seasonal" = cpi_seasonal_adjustment_fy,
+             "trimmed.mean" = cpi_trimmed_fy)
+    }
+  
   
   permitted_fys <- .subset2(cpi.indices, "fy_year")
   earliest_from_fy <- permitted_fys[[1L]]
@@ -227,3 +236,5 @@ cpi_inflator <- function(from_nominal_price = 1,
            time.col = "fy_year")
   
 }
+
+