+++ tests/testthat/test_wage_inflator.R
@@ -153,5 +153,13 @@ test_that("accelerated", {
   expect_identical(aff, bff)
 })
 
+test_that("accelerating both from and to", {
+  expect_identical(wage_inflator(from_fy = c("2005-06", "2008-09", "2006-07"),
+                                 to_fy = c("2015-16", "2014-15", "2016-17")),
+                   wage_inflator(from_fy = c("2005-06", "2008-09", "2006-07"),
+                                 to_fy = c("2015-16", "2014-15", "2016-17"),
+                                 accelerate.above = 2L))
+})
+