+++ NEWS.md
@@ -11,9 +11,10 @@
 
 Bug fixes:
 * `inflator` no longer fails when `to_fy` is length > 1 and unordered.
-* `inflator` and `cpi_inflator` is now much faster:
+* `inflator` and `cpi_inflator`, `lf_inflator_fy`, and `wage_inflator` are now much faster when either `from_fy` or `to_fy` have more than 100,000 elements:
 
 ```r
+set.seed(19952010)
 from_fys <- sample(yr2fy(1995:2010), size = 1e6, replace = TRUE)
 microbenchmark(cpi_inflator(from_fy = from_fys, to_fy = "2015-16"))
 # Old